+++ CHANGELOG.md
@@ -17,6 +17,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Rename `ReplicationRules` into `ReplicationFns` with its module.
 - Rename `ReplicationInfo` into `ComponentFns`.
+- Rename `DespawnEntityFn` into `DespawnFn`.
+- Rename `RemoveComponentFn` into `RemoveFn`.
 - Rename `despawn_fn` field from `ReplicationFns` into `despawn`.
 - Move `AppReplicationExt` and default component functions into `component_rules` module.
 - `scene::replicate_into` now relies on the newly added `ReplicatedArchetypes` that updated in `ServerSet::UpdateArchetypes`. So it's preferred to replicate into scene after this set.